fix(rs): use u64 for pairwise product to prevent overflow on 32-bit
On 32-bit targets, `n * m` in `usize` overflows when both sample
sizes exceed ~65536. Use `u64` for the product and downstream
arithmetic in pairwise_margin and shift_bounds. Also apply rustfmt
to estimators.rs.

+++ rs/pragmastat/src/estimators.rs
@@ -12,8 +12,7 @@ use crate::assumptions::{
 pub fn center(x: &[f64]) -> Result<f64, EstimatorError> {
     // Check validity (priority 0)
     check_validity(x, Subject::X)?;
-    crate::fast_center::fast_center(x)
-        .map_err(EstimatorError::from)
+    crate::fast_center::fast_center(x).map_err(EstimatorError::from)
 }
 
 /// Estimates data dispersion (spread)
@@ -35,8 +34,7 @@ pub fn spread(x: &[f64]) -> Result<f64, EstimatorError> {
     // Check validity first (priority 0)
     check_validity(x, Subject::X)?;
 
-    let spread_val = crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?;
+    let spread_val = crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)?;
     if spread_val <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
@@ -66,14 +64,12 @@ pub fn rel_spread(x: &[f64]) -> Result<f64, EstimatorError> {
     check_positivity(x, Subject::X)?;
 
     // Calculate center (we know x is valid, center should succeed)
-    let center_val = crate::fast_center::fast_center(x)
-        .map_err(EstimatorError::from)?;
+    let center_val = crate::fast_center::fast_center(x).map_err(EstimatorError::from)?;
 
     // Calculate spread (we know x is valid)
     // Note: spread now requires sparity, but for rel_spread we require positivity
     // which is checked above. We use the internal implementation directly.
-    let spread_val = crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?;
+    let spread_val = crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)?;
 
     // center_val is guaranteed positive because all values are positive
     Ok(spread_val / center_val.abs())
@@ -88,8 +84,7 @@ pub fn shift(x: &[f64], y: &[f64]) -> Result<f64, EstimatorError> {
     // Check validity (priority 0)
     check_validity(x, Subject::X)?;
     check_validity(y, Subject::Y)?;
-    crate::fast_shift::fast_shift(x, y)
-        .map_err(EstimatorError::from)
+    crate::fast_shift::fast_shift(x, y).map_err(EstimatorError::from)
 }
 
 /// Measures how many times larger x is compared to y (ratio)
@@ -122,8 +117,7 @@ pub fn ratio(x: &[f64], y: &[f64]) -> Result<f64, EstimatorError> {
     // Check positivity for y (priority 1, subject y)
     check_positivity(y, Subject::Y)?;
 
-    crate::fast_shift::fast_ratio(x, y)
-        .map_err(EstimatorError::from)
+    crate::fast_shift::fast_ratio(x, y).map_err(EstimatorError::from)
 }
 
 /// Measures the typical variability when considering both samples together (avg_spread)
@@ -151,13 +145,11 @@ pub(crate) fn avg_spread(x: &[f64], y: &[f64]) -> Result<f64, EstimatorError> {
     let n = x.len();
     let m = y.len();
 
-    let spread_x = crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?;
+    let spread_x = crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)?;
     if spread_x <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
-    let spread_y = crate::fast_spread::fast_spread(y)
-        .map_err(EstimatorError::from)?;
+    let spread_y = crate::fast_spread::fast_spread(y).map_err(EstimatorError::from)?;
     if spread_y <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::Y)));
     }
@@ -189,20 +181,17 @@ pub fn disparity(x: &[f64], y: &[f64]) -> Result<f64, EstimatorError> {
     let n = x.len();
     let m = y.len();
 
-    let spread_x = crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?;
+    let spread_x = crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)?;
     if spread_x <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
-    let spread_y = crate::fast_spread::fast_spread(y)
-        .map_err(EstimatorError::from)?;
+    let spread_y = crate::fast_spread::fast_spread(y).map_err(EstimatorError::from)?;
     if spread_y <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::Y)));
     }
 
     // Calculate shift (we know inputs are valid)
-    let shift_val = crate::fast_shift::fast_shift(x, y)
-        .map_err(EstimatorError::from)?;
+    let shift_val = crate::fast_shift::fast_shift(x, y).map_err(EstimatorError::from)?;
 
     let avg_spread_val = (n as f64 * spread_x + m as f64 * spread_y) / (n + m) as f64;
 
@@ -263,7 +252,7 @@ pub fn shift_bounds(x: &[f64], y: &[f64], misrate: f64) -> Result<Bounds, Estima
     xs.sort_by(|a, b| a.total_cmp(b));
     ys.sort_by(|a, b| a.total_cmp(b));
 
-    let total = n * m;
+    let total = n as u64 * m as u64;
 
     // Special case: when there's only one pairwise difference, bounds collapse to a single value
     if total == 1 {
@@ -530,16 +519,10 @@ fn avg_spread_bounds_with_rngs(
         )));
     }
 
-    if crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?
-        <= 0.0
-    {
+    if crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)? <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
-    if crate::fast_spread::fast_spread(y)
-        .map_err(EstimatorError::from)?
-        <= 0.0
-    {
+    if crate::fast_spread::fast_spread(y).map_err(EstimatorError::from)? <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::Y)));
     }
 
@@ -644,16 +627,10 @@ fn disparity_bounds_with_rngs(
     let alpha_shift = min_shift + extra / 2.0;
     let alpha_avg = min_avg + extra / 2.0;
 
-    if crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?
-        <= 0.0
-    {
+    if crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)? <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
-    if crate::fast_spread::fast_spread(y)
-        .map_err(EstimatorError::from)?
-        <= 0.0
-    {
+    if crate::fast_spread::fast_spread(y).map_err(EstimatorError::from)? <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::Y)));
     }
 
@@ -762,10 +739,7 @@ fn spread_bounds_with_rng(
     if x.len() < 2 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
-    if crate::fast_spread::fast_spread(x)
-        .map_err(EstimatorError::from)?
-        <= 0.0
-    {
+    if crate::fast_spread::fast_spread(x).map_err(EstimatorError::from)? <= 0.0 {
         return Err(EstimatorError::from(AssumptionError::sparity(Subject::X)));
     }
